note authorization set up
note creation is now working fully minus the ability to stop duplicates from being made. I need to add a unique id to each note to check for duplicates and also to use as a unique key to pull and delete from the table when needed. The note authorization page has also been completed. The application pulls the info from the oldest note saved in new notes and displays it on the screen. the authorizer looks over the note and either signs it or writes a comment. if the not is signed it goes to the saved notes tables and is deleted from new notes, else it stays in new notes and an email is sent to the writer with an explanation of what needs to be changed.