lvl restriciton, unique note ids, revision tab
Set up the level restrictions for the login. Now only the people with the correct restrictions can access authorization of notes and creating new users. Finished the note authorization. Now there are unique ids to go with each note created. The notes can be searched and deleted via these ids and in the emails sending to the authors of the notes if they are incomplete, there is the note id accompanying it. I also formatted the note id's so that it is specific to each person. It says their name and what needs to be fixed. I also added a not revision tab so that if you are emailed about a bad note, you go there to enter in the note id and pull up the note that is wrong.